car-rent commit 45 APK fix long id

+++ app/src/main/java/com/example/composecarrent/ui/theme/admin_panel/AdminPanelScreenBody.kt
@@ -77,7 +77,7 @@ fun AdminPanelScreenBody(
     val carIcon3 = viewModel.photoUri3?.let { viewModel.imageToBase64(it, contentResolver) }
 
     val newCar = CarDataModel(
-        carId.toInt(),
+        carId,
         category,
         carIcon1 ?: "",
         carIcon2 ?: "",
